Working on adding update item and order to the views

+++ src/hud/htmx_views.py
@@ -1,12 +1,12 @@
+from decimal import Decimal
 from django.http import JsonResponse
 from django.shortcuts import get_object_or_404, render
 from src.hud.models import ProductGroup, Product, Barcode, PosOrderItem, PosOrder
-from decimal import Decimal
-
-active_order = PosOrder.objects.filter(is_active=True).first()
+from src.hud.calculations import add_or_update_product_to_order
 
 
 def modal_product(request, id):
+    active_order = PosOrder.objects.filter(is_active=True).first()
     product = get_object_or_404(Product, id=id)
 
     if product:
@@ -15,6 +15,7 @@ def modal_product(request, id):
 
 
 def add_quantity(request, item_number):
+    active_order = PosOrder.objects.filter(is_active=True).first()
     item = get_object_or_404(PosOrderItem, number=item_number)
     item.quantity += 1  # Set quantity to the new value received from the client
     item.save()
@@ -27,6 +28,7 @@ def add_quantity(request, item_number):
 
 
 def subtract_quantity(request, item_number):
+    active_order = PosOrder.objects.filter(is_active=True).first()
     item = get_object_or_404(PosOrderItem, number=item_number)
     context = {
         "item": item,
@@ -37,10 +39,23 @@ def subtract_quantity(request, item_number):
         item.save()
         return render(request, 'hud/pos/renders/update-order-item.html', context)
     elif item.quantity == 1:
-        return render(request, 'hud/pos/renders/confirm-remove.html', context)
+        return render(request, 'hud/pos/renders/order-item-with-confirm.html', context)
 
 
-def remove_item(request, item_number):
+def remove_item(request, item_number, is_button=False):
+    '''
+    Check if it is a direct delete
+    '''
+    if is_button:
+        active_order = PosOrder.objects.filter(is_active=True).first()
+        item = get_object_or_404(PosOrderItem, number=item_number)
+        context = {
+            "item": item,
+            "active_order": active_order,
+        }
+        return render(request, 'hud/pos/renders/order-item-with-confirm.html', context)
+
+    active_order = PosOrder.objects.filter(is_active=True).first()
     item = get_object_or_404(PosOrderItem, number=item_number)
     item.delete()
 
@@ -52,75 +67,57 @@ def remove_item(request, item_number):
 
 def add_item_with_barcode(request):
 
+    active_order = PosOrder.objects.filter(is_active=True).first()
     barcode_value = request.POST.get("barcode", None)
-    print(barcode_value)
     barcode = get_object_or_404(Barcode, value=barcode_value)
+    order = active_order
+    total = 0
     item = PosOrderItem.objects.filter(
         order=active_order, product=barcode.product).first()
 
     if not item:
-        item = PosOrderItem.objects.create(
-            user=request.user,
-            order=active_order,
-            product=barcode.product,
-            price=barcode.product.price,
-            quantity=1
-        )
+        order, item, total = add_or_update_product_to_order(
+            request.user, barcode.product, active_order)
     else:
         item.quantity += 1
         item.save()
 
     context = {
         "item": item,
-        "active_order": active_order,
+        "active_order": order,
+        "total": total
     }
     return render(request, 'hud/pos/renders/update-active-order.html', context)
 
 
 def add_order_item(request):
+    active_order = PosOrder.objects.filter(is_active=True).first()
     if request.method == 'POST':
         product_id = request.POST.get('product_id', None)
-        quantity_str = request.POST.get('quantity', '1')
-        try:
-            quantity = Decimal(quantity_str)
-        except Decimal.InvalidOperation as e:
-            print(e)
-            # Handle the case where quantity is not a valid decimal
-            return
+        quantity = request.POST.get('quantity', 1)
 
-        active_order = PosOrder.objects.filter(is_active=True).first()
+        print("Quantity1 = ", quantity)
 
-        if product_id:
-            product = get_object_or_404(Product, id=product_id)
-            item = PosOrderItem.objects.filter(product=product).first()
-            print("type(product.price) = ", type(product.price))
-            if item:
-                context = {
-                    "item": item,
-                    "active_order": active_order,
-                }
-                if quantity:
-                    item.quantity += quantity
+        order = active_order
 
-                    item.save()
-                    return render(request, 'hud/pos/renders/update-active-order.html', context)
-            else:
-                item = PosOrderItem.objects.create(
-                    user=request.user,
-                    order=active_order,
-                    product=product,
-                    price=product.price
-                    # quantity=int(quantity)
-                )
-                if quantity > 1:
-                    item.quantity = quantity
-                    item.save()
-                context = {
-                    "item": item,
-                    "active_order": active_order,
-                }
-            return render(request, 'hud/pos/renders/update-active-order.html', context)
+        product = get_object_or_404(Product, id=product_id)
+
+        item = PosOrderItem.objects.filter(
+            order=active_order, product=product).first()
+
+        if not item:
+            order, subtotal, item = add_or_update_product_to_order(
+                request.user, product, quantity, active_order)
         else:
-            return JsonResponse({"failed": "Failed to add!"})
+            item.quantity += 1
+            item.save()
+
+        context = {
+            "item": item,
+            "items_subtotal": subtotal,
+            "active_order": order,
+        }
+
+        return render(request, 'hud/pos/renders/update-active-order.html', context)
 
     return JsonResponse({"success": "Item added successfully!"})
